refactor: 페이지에 ui 폴더 생성
