+++ client/src/pages/MainPage/index.ts
@@ -1,1 +1,1 @@
-export { MainPage } from './MainPage';
+export { MainPage } from './ui/MainPage';
