+++ client/src/pages/MainPage/ui/MainPage.tsx
