+++ client/src/pages/ManagePage/index.ts
@@ -1,1 +1,1 @@
-export { ManagePage } from './ManagePage';
+export { ManagePage } from './ui/ManagePage';
