+++ client/src/pages/StreamingPage/ui/StreamingPage.tsx
@@ -1,7 +1,7 @@
 import { ChattingContainer } from '@/widgets/chatting';
 import { Vote } from '@/widgets/vote';
 import { Streaming } from '@/widgets/streaming';
-import { useStreamingRoom } from './useStreamingRoom';
+import { useStreamingRoom } from '@/shared/hook/useStreamingRoom';
 export function StreamingPage() {
   // TODO: 연결 여부 전역 상태로 관리하기
   // TODO: 소켓 이벤트들 handler로 만들어서 한번에 관리하기
